add end_auction instruction + tests

+++ program/programs/nft-marketplace/src/state/user.rs
@@ -3,9 +3,9 @@ use anchor_lang::prelude::*;
 #[account]
 #[derive(InitSpace)]
 pub struct UserAccount {
-    pub total_bids_placed: u64,
-    pub total_auctions_participated: u64,
-    pub total_auctions_won: u64,
+    pub total_bids_placed: u32,
+    pub total_auctions_participated: u32,
+    pub total_auctions_won: u32,
 
     pub points: u32,
     pub bump: u8,
